Enhance button decorations and added A/C function
Implemented visual dynamics to buttons so they light up when clicked. Set
a limit on the numbers on display to only 10 digits and made a clear button
to quickly reset the display screen. Having trouble with the operators ignored
figuring out how to store two numbers to use, will look into it next time.