Added decimal operator
A little more tricky to implement. Modified default streamLine value to be 0
now so that any operation done will first check if it is 0 already and adjust
accordingly. Main concern was to check if number on screen was already a decimal.
If it is, don't have to add any more decimal dots. Caught an error of using
streamLine.includes() for the decimal dot and the input type was function, have to
explicitly typecast to String. Still have the float precision problem to
address (1.63 + 0.47 = 2.099996)