Added signs button for +/- and fixed edge cases
Minor changes to visual output for more edge cases of the number 0. Also
implemented the sign conversion for +/- button, which required type conversions.
Furthermore, accounted for the test case of small decimal numbers with many
decimal digits via rounding, which fixed the previous issues of floating inaccuracies
for small numbers. However, very large decimal numbers continue to be a
problem with staying within the calculator screen.