Add webpage interaction with visual btn responses
Implemented a class toggle to the operation buttons so that they visually
respond and stay that way when clicked. The code did not work at first, but
found out about CSS class chaining priority, which was leading to why the updated
decorations did not appear. As such, increased .active class specificty through
chaining, which helped the CSS modifications take priority over others.